requirements.txt
flask
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
groq
python-dotenv
pypdf
gunicorn